added test for EntityManagersHandler and fixed some static analysis problems
