+++ composer.json
@@ -32,7 +32,7 @@
         "doctrine/annotations": "^1.6",
         "doctrine/doctrine-bundle": "^1.10",
         "doctrine/doctrine-migrations-bundle": "^2.0",
-        "doctrine/orm": "^2.6.3"
+        "doctrine/orm": "^2.6.3",
         "friendsofphp/php-cs-fixer": "^2.14",
         "jangregor/phpstan-prophecy": "^0.3.0",
         "phpstan/phpstan": "^0.11.4",
